# docs/api/openapi.yaml
openapi: 3.2.0
...
info:
  title: Imperio Industrial — API pública
  version: 1.5.0
  summary: API REST del MMO de simulación económica, industrial y logística "Imperio Industrial".
  description: |
    API pública del servidor autoritativo de **Imperio Industrial** (GDD/SAD v1.2).

    ## Principios del contrato de API

    - **Una sola API para humanos y bots** (igualdad de API literal, ADR-010): mismos
      endpoints y mismos rate limits lógicos. El motor no distingue el origen de un comando.
    - **Servidor autoritativo**: el cliente solo envía intenciones; toda validación
      (espacio, fondos, stock, plazos) ocurre server-side.
    - **REST para operaciones no urgentes**; los eventos en tiempo real del área de interés
      (movimiento de vehículos, alertas configuradas) se distribuyen por **WebSocket**
      a través del Notification/Event Gateway y quedan fuera de este documento.
    - **El tablón global es pull, no push**: se consulta bajo demanda con filtros
      (producto, ubicación, precio, plazo); nunca hay push mundial de mercado.
    - **Sim-time en el contrato**: todo plazo de dominio (contratos, producción, viajes)
      se define y transmite en sim-time (segundos desde el génesis del mundo, ratio 24×).
      La traducción a tiempo real es responsabilidad exclusiva de la UI. Las dos únicas
      mecánicas en tiempo real son la ventana de sorteo y el cooldown anti-parpadeo.
...
      (invariante del ledger).
...
      distinguibles.
...

    ## Estructura de respuestas

...

...
servers:
...
security:
  - sessionToken: []
...
paths:
...
  /world/vehicle-types:
    get:
      operationId: listVehicleTypes
      tags: [world, fleet]
      summary: Catálogo de tipos de vehículo
      parameters:
        - name: mode
          in: query
          schema:
            $ref: '#/components/schemas/LinkMode'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Tipos de vehículo (camión/tren/barco; el aéreo es expansión futura).
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VehicleType'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/vehicles:
    get:
      operationId: listVehicles
      tags: [world, fleet]
      summary: Flota propia
      description: Paneles agregados de flota; la posición de vehículos en tránsito se deriva analíticamente al observarla.
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VehicleStatus'
        - name: route_id
          in: query
          schema:
            $ref: '#/components/schemas/RouteId'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Vehículos de la corporación.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vehicle'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
    post:
      operationId: purchaseVehicle
      tags: [world, fleet]
      summary: Comprar un vehículo
      description: Compra a precio de catálogo con entrega en un nodo propio o terminal accesible. El mercado secundario entre jugadores usa el CCRI estándar.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehiclePurchase'
      responses:
        '201':
          description: Vehículo adquirido (`idle` en el nodo indicado).
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Vehicle'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: Fondos insuficientes (`INSUFFICIENT_FUNDS`) o nodo de entrega incompatible con el modo del vehículo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/vehicles/{vehicleId}:
    parameters:
      - name: vehicleId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/VehicleId'
    get:
      operationId: getVehicle
      tags: [world, fleet]
      summary: Detalle de un vehículo
      description: Incluye la posición derivada analíticamente en el momento de la consulta.
      responses:
        '200':
          description: Vehículo.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Vehicle'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
    patch:
      operationId: updateVehicle
      tags: [world, fleet]
      summary: Comandar un vehículo
      description: |
        Asigna/retira una ruta o programa mantenimiento. Un vehículo `sealed`
        (SELLADO durante un handoff entre shards) es visible pero **no comandable**:
        responde 403 `VEHICLE_SEALED` (solo aplica tras la extracción multi-proceso).
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdate'
      responses:
        '200':
          description: Vehículo actualizado.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Vehicle'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: 'El vehículo pertenece a otra corporación, o está SELLADO durante un handoff (`VEHICLE_SEALED`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
...
  /world/shipments:
    get:
      operationId: listShipments
      tags: [world, fleet]
      summary: Cargamentos visibles (propios y los que se transportan por flete)
      description: |
        Cargamentos etiquetados por contrato: el stock reservado viaja sin dejar de estar
        reservado. **Nada se teletransporta, tampoco en los fallos** — el stock de un
        contrato fallido se libera en su ubicación física actual (`released_in_situ`).

        Devuelve los cargamentos **propios** y, en un **CCRI-Flete**, los que corresponden
        a la corporación como **transportista**: el dueño del cargamento es el cargador,
        pero quien lo despacha y lo lleva es el transportista (GDD 5.3.2), así que necesita
        verlo. Misma regla de autorización que `POST /world/shipments/{shipmentId}/dispatch`.
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ShipmentStatus'
        - name: contract_id
          in: query
          schema:
            $ref: '#/components/schemas/ContractId'
        - name: freight_contract_id
          in: query
          description: Filtra por el CCRI-Flete del cargamento (el que un transportista debe despachar).
          schema:
            $ref: '#/components/schemas/FreightContractId'
        - name: vehicle_id
          in: query
          schema:
            $ref: '#/components/schemas/VehicleId'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Cargamentos.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Shipment'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
...
  /world/terminals/{terminalId}:
    get:
      operationId: getTerminal
      tags: [world]
      summary: Detalle de una terminal
      description: Las terminales tienen dueño y cola de transbordo; el gameplay de "infraestructura como servicio" vive en los nodos.
      parameters:
        - name: terminalId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TerminalId'
      responses:
        '200':
          description: Terminal.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Terminal'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        Token de sesión emitido por `POST /auth/sessions`. El servidor solo almacena su
        hash. Humanos y bots usan el mismo esquema (credenciales por cuenta); los bots
        acceden por un camino de red interno multiplexado con la misma superficie de API.
...